+++ src/core/response_wire.c
@@ -64,7 +64,15 @@ struct response_wire_s {
         size_t   ct_off,    ct_len;
         size_t   dn_off,    dn_len;
         size_t   cc_off,    cc_len;
+        size_t   mth_off,   mth_len;
+        size_t   uri_off,   uri_len;
+        size_t   rng_off,   rng_len;
+        size_t   ifr_off,   ifr_len;
+        size_t   ims_off,   ims_len;
+        size_t   inm_off,   inm_len;
         bool     ct_present, dn_present, cc_present;
+        bool     mth_present, uri_present, rng_present;
+        bool     ifr_present, ims_present, inm_present;
         int      status;
         uint8_t  disposition;
         bool     disposition_set, etag, last_modified, accept_ranges,
@@ -189,6 +197,12 @@ bool response_wire_set_send_file(response_wire_t *rw,
     WIRE_SF_STR(content_type,  ct_off, ct_len, ct_present)
     WIRE_SF_STR(download_name, dn_off, dn_len, dn_present)
     WIRE_SF_STR(cache_control, cc_off, cc_len, cc_present)
+    WIRE_SF_STR(method,        mth_off, mth_len, mth_present)
+    WIRE_SF_STR(uri,           uri_off, uri_len, uri_present)
+    WIRE_SF_STR(range,         rng_off, rng_len, rng_present)
+    WIRE_SF_STR(if_range,      ifr_off, ifr_len, ifr_present)
+    WIRE_SF_STR(if_modified_since, ims_off, ims_len, ims_present)
+    WIRE_SF_STR(if_none_match, inm_off, inm_len, inm_present)
 #undef WIRE_SF_STR
 
     rw->sf.status            = sf->status;
@@ -220,6 +234,19 @@ bool response_wire_get_send_file(const response_wire_t *rw,
     out->download_name_len = rw->sf.dn_present ? rw->sf.dn_len : 0;
     out->cache_control     = rw->sf.cc_present ? rw->arena + rw->sf.cc_off : NULL;
     out->cache_control_len = rw->sf.cc_present ? rw->sf.cc_len : 0;
+
+#define WIRE_SF_GET(field, off_m, len_m, present_m)                          \
+    out->field       = rw->sf.present_m ? rw->arena + rw->sf.off_m : NULL;   \
+    out->field##_len = rw->sf.present_m ? rw->sf.len_m : 0;
+
+    WIRE_SF_GET(method,            mth_off, mth_len, mth_present)
+    WIRE_SF_GET(uri,               uri_off, uri_len, uri_present)
+    WIRE_SF_GET(range,             rng_off, rng_len, rng_present)
+    WIRE_SF_GET(if_range,          ifr_off, ifr_len, ifr_present)
+    WIRE_SF_GET(if_modified_since, ims_off, ims_len, ims_present)
+    WIRE_SF_GET(if_none_match,     inm_off, inm_len, inm_present)
+#undef WIRE_SF_GET
+
     out->status            = rw->sf.status;
     out->disposition       = rw->sf.disposition;
     out->disposition_set   = rw->sf.disposition_set;